XmlRpcPlugin: Implemented a test subsystem, and the start of functional unittests - #5382.
Test subsystem depends on a patch for Trac 0.11-stable (not yet committed - see trac:ticket:8388), and Trac needs to be running from source (either installing Trac using `python setup.py develop`, or adding Trac source to `PYTHONPATH` when running tests - it reuses Trac functional test infrastructure to create a test project and start a standalone server that answer requests for the duration of the test run.

The test server is available with basic authentication and 3 users (anonymous|user|admin), and any aspect of the plugin should be testable using this infrastructure. Currently it contains just a couple of complicated tests for workflow and security policy plugins, but a long list of plain method->result tests also needs to be added. Patches welcome :-)

Running tests for plugin: `python setup.py test`